+++ src/org/ojalgo/function/aggregator/AggregatorFunction.java
@@ -27,13 +27,20 @@ import org.ojalgo.structure.AccessScalar;
 
 public interface AggregatorFunction<N extends Number> extends VoidFunction<N>, AccessScalar<N> {
 
+    /**
+     * @deprecated v48 Merging will no longer be supported
+     */
+    @Deprecated
     default boolean isMergeable() {
         return true;
     }
 
     /**
      * Only works if {@link #isMergeable()}!
+     *
+     * @deprecated v48 Merging will no longer be supported
      */
+    @Deprecated
     void merge(N result);
 
     AggregatorFunction<N> reset();